fixed random rendering
