+++ clients/apple/Shared/Views/BookView.swift
@@ -62,7 +62,7 @@ struct BookView: View {
                     ToolbarItemGroup {
                         NavigationLink(
                             destination: PendingSharesView()) {
-                                Image(systemName: "shared.with.you").foregroundColor(.blue)
+                                Image(systemName: "person.3.fill").foregroundColor(.blue)
                             }
                         
                         NavigationLink(
