+++ clients/apple/Shared/Views/ListFileCell.swift
@@ -76,198 +76,3 @@ struct RealFileCell: View {
             .contentShape(Rectangle()) /// https://stackoverflow.com/questions/57258371/swiftui-increase-tap-drag-area-for-user-interaction
     }
 }
-
-struct SearchFilePathCell: View {
-    let name: String
-    let path: String
-    let matchedIndices: [Int]
-    
-    @State var nameModified: Text
-    @State var pathModified: Text
-    @State var modifiedGenerated: Bool
-    
-    init(name: String, path: String, matchedIndices: [Int]) {
-        self.name = name
-        self.path = path
-        self.matchedIndices = matchedIndices
-        
-        self._modifiedGenerated = State(initialValue: false)
-        self._nameModified = State(initialValue: Text(name))
-        self._pathModified = State(initialValue: Text(path))
-    }
-    
-    @Environment(\.horizontalSizeClass) var horizontal
-    @Environment(\.verticalSizeClass) var vertical
-    
-    var body: some View {
-        VStack(alignment: .leading, spacing: 5) {
-            nameModified
-                .font(.title3)
-            
-            HStack {
-                Image(systemName: "doc")
-                    .foregroundColor(.accentColor)
-                    .font(.caption)
-                
-                pathModified
-                        .foregroundColor(.accentColor)
-                        .font(.caption)
-            }
-        }
-            .contentShape(Rectangle()) /// https://stackoverflow.com/questions/57258371/swiftui-increase-tap-drag-area-for-user-interaction
-            .iOSSearchPadding(horizontal: horizontal, vertical: vertical)
-            .onAppear {
-                underlineMatchedSegments()
-            }
-    }
-    
-    func underlineMatchedSegments() {
-        if(modifiedGenerated) {
-            return
-        }
-        
-        modifiedGenerated = true
-        
-        let matchedIndicesHash = Set(matchedIndices)
-        
-        var pathOffset = 1;
-        
-        if(path.count - 1 > 0) {
-            pathModified = Text("")
-            
-            for index in 0...path.count - 1 {
-                let correctIndex = String.Index(utf16Offset: index, in: path)
-                let newPart = Text(path[correctIndex...correctIndex])
-                                
-                if(path[correctIndex...correctIndex] == "/") {
-                    pathModified = pathModified + Text(" > ").foregroundColor(.gray)
-                } else if(matchedIndicesHash.contains(index + 1)) {
-                    pathModified = pathModified + newPart.bold()
-                } else {
-                    pathModified = pathModified + newPart
-                }
-            }
-            
-            pathOffset = 2
-        }
-                
-        if(name.count - 1 > 0) {
-            nameModified = Text("")
-            for index in 0...name.count - 1 {
-                let correctIndex = String.Index(utf16Offset: index, in: name)
-                let newPart = Text(name[correctIndex...correctIndex])
-                
-                if(matchedIndicesHash.contains(index + path.count + pathOffset)) {
-                    nameModified = nameModified + newPart.bold()
-                } else {
-                    nameModified = nameModified + newPart.foregroundColor(.gray)
-                }
-            }
-        }
-    }
-}
-
-struct SearchFileContentCell: View {
-    let name: String
-    let path: String
-    let paragraph: String
-    let matchedIndices: [Int]
-    
-    @State var paragraphModified: Text
-    @State var pathModified: Text
-    @State var modifiedGenerated: Bool
-    
-    init(name: String, path: String, paragraph: String, matchedIndices: [Int]) {
-        self.name = name
-        self.path = path
-        self.paragraph = paragraph
-        self.matchedIndices = matchedIndices
-        
-        self._modifiedGenerated = State(initialValue: false)
-        self._pathModified = State(initialValue: Text(path))
-        self._paragraphModified = State(initialValue: Text(paragraph))
-    }
-    
-    @Environment(\.horizontalSizeClass) var horizontal
-    @Environment(\.verticalSizeClass) var vertical
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 5) {
-            Text(name)
-                .font(.title3)
-                .foregroundColor(.gray)
-            
-            HStack {
-                Image(systemName: "doc")
-                    .foregroundColor(.accentColor)
-                    .font(.caption2)
-                
-                pathModified
-                        .foregroundColor(.accentColor)
-                        .font(.caption2)
-            }
-            .padding(.bottom)
-            
-            paragraphModified
-                .font(.caption)
-        }
-            .contentShape(Rectangle()) /// https://stackoverflow.com/questions/57258371/swiftui-increase-tap-drag-area-for-user-interaction
-            .iOSSearchPadding(horizontal: horizontal, vertical: vertical)
-            .onAppear {
-                underlineMatchedSegments()
-            }
-    }
-    
-    func underlineMatchedSegments() {
-        if(modifiedGenerated) {
-            return
-        }
-        
-        modifiedGenerated = true
-
-        let matchedIndicesHash = Set(matchedIndices)
-        
-        if(path.count - 1 > 0) {
-            pathModified = Text("")
-            
-            for index in 0...path.count - 1 {
-                let correctIndex = String.Index(utf16Offset: index, in: path)
-                                
-                if(path[correctIndex...correctIndex] == "/") {
-                    pathModified = pathModified + Text(" > ").foregroundColor(.gray)
-                } else {
-                    pathModified = pathModified + Text(path[correctIndex...correctIndex])
-                }
-            }
-        }
-                
-        if(paragraph.count - 1 > 0) {
-            paragraphModified = Text("")
-            
-            for index in 0...paragraph.count - 1 {
-                let correctIndex = String.Index(utf16Offset: index, in: paragraph)
-                let newPart = Text(paragraph[correctIndex...correctIndex])
-                
-                if(matchedIndicesHash.contains(index)) {
-                    paragraphModified = paragraphModified + newPart.bold()
-                } else {
-                    paragraphModified = paragraphModified + newPart.foregroundColor(.gray)
-                }
-            }
-            
-        }
-    }
-}
-
-extension View {
-    public func iOSSearchPadding(horizontal: UserInterfaceSizeClass?, vertical: UserInterfaceSizeClass?) -> some View {
-        Group {
-            if horizontal == .regular && vertical == .regular {
-                self
-            } else {
-                self
-                    .padding(.vertical, 5)
-            }
-        }
-    }
-}
